Refactored the validateRecordExists method inside the validator-middleware.ts file
